+++ src/core/entities/order.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass, field
+from decimal import Decimal
 from uuid import UUID
 
 from src.core.enums import OrderStatus
@@ -8,6 +9,6 @@ from src.core.enums import OrderStatus
 class Order:
     id: UUID
     status: OrderStatus
-    total_price: float
+    total_price: Decimal
     user_id: UUID
     product_ids: list[UUID] = field(default_factory=list)
