+++ src/core/entities/order_detail.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass, field
 from datetime import datetime
+from decimal import Decimal
 from uuid import UUID
 
 from src.core.enums import OrderStatus
@@ -10,6 +11,6 @@ from src.core.entities.order_item import OrderItem
 class OrderDetail:
     id: UUID
     status: OrderStatus
-    total_price: float
+    total_price: Decimal
     created_at: datetime
     items: list[OrderItem] = field(default_factory=list)
